处理数据库Production的功率时间带了单位的问题

+++ define.h
@@ -228,8 +228,8 @@ struct _Production_Data
     int pressure;                          // 压力
     QString time;                          // 焊接时间
     int power;                             // 功率
-    int pre_height;                        // 焊前高度
-    int post_height;                       // 焊后高度
+    double pre_height;                        // 焊前高度
+    double post_height;                       // 焊后高度
     int force;                             // 撕拉力
     int residual;                          // 残留度
     int good_rate;                         // 良率
